correcting files renaming
