+++ src/app/gifts/gifts/gifts.component.ts
@@ -3,8 +3,6 @@ import { Component } from '@angular/core';
 @Component({
   selector: 'app-gifts',
   templateUrl: './gifts.component.html',
-  styleUrls: ['./gifts.component.scss']
+  styleUrls: ['./gifts.component.scss'],
 })
-export class GiftsComponent {
-
-}
+export class GiftsPageComponent {}
