+++ src/app/requests/requests.module.ts
@@ -4,16 +4,16 @@ import { AppRoutingModule } from '../app-routing.module';
 import { RouterOutlet } from '@angular/router';
 import { MatTabsModule } from '@angular/material/tabs';
 import { SharedModule } from 'src/shared/libs/shared.module';
-import { RequestsSectionComponent } from './requests.component';
-import { RequestsComponent } from './requests/requests.component';
+import { RequestsComponent } from './requests.component';
+import { RequestsPageComponent } from './requests/requests.component';
 import { ComplaintsComponent } from './complaints/complaints.component';
 import { ClosedComplaintsComponent } from './closed-complaints/closed-complaints.component';
 import { ClosedRequestsComponent } from './closed-requests/closed-requests.component';
 
 @NgModule({
   declarations: [
-    RequestsSectionComponent,
     RequestsComponent,
+    RequestsPageComponent,
     ComplaintsComponent,
     ClosedComplaintsComponent,
     ClosedRequestsComponent,
