--- src/app/settings/distributers/distributers.component.ts
@@ -1,10 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-distributers',
-  templateUrl: './distributers.component.html',
-  styleUrls: ['./distributers.component.scss']
-})
-export class DistributersComponent {
-
-}
