+++ src/app/visits/distributer/distributer.component.ts
@@ -1,0 +1,8 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-distributer',
+  templateUrl: './distributer.component.html',
+  styleUrls: ['./distributer.component.scss'],
+})
+export class DistributerComponent {}
